Played a bit with the Pixel Editor, read descr.
I need your help, fellow GitHub followers/readers.
I need a 2012 (possibly June-July) Pixel Editor client, since the actual
one has been Kabam'd (basically, it requires their SSL certificate - I'm
working on editing the client but it's really hard for me since I have
no knowledge at all in AS3/bytecode) to get it to work without too many
problems.

+++ wServer/ClientProcessor.cs
@@ -149,7 +149,7 @@ namespace wServer
                     Console.WriteLine("Unhandled packet: " + pkt.ToString());
                 }
                 else
-                    Console.WriteLine(pkt.ID);
+                    Console.WriteLine((int)pkt.ID);
             }
             catch
             {
